requirements.txt
Django==4.0.2
djangorestframework==3.13.1
pytest-django==4.5.2
freezegun==1.1.0